frameworks-comparison/benchmark/requirements.txt
dbos
temporalio
prometheus_client
httpx
resonate-sdk
